feedback + refactoring of pthreads folder
