+++ src/mono/browser/runtime/diagnostics/browser/controller.ts
@@ -7,7 +7,7 @@ import { threads_c_functions as cwraps } from "../../cwraps";
 import { INTERNAL, mono_assert } from "../../globals";
 import { mono_log_info, mono_log_debug, mono_log_warn } from "../../logging";
 import { withStackAlloc, getI32 } from "../../memory";
-import { waitForThread } from "../../pthreads/browser";
+import { waitForThread } from "../../pthreads";
 import { isDiagnosticMessage, makeDiagnosticServerControlCommand } from "../shared/controller-commands";
 import monoDiagnosticsMock from "consts:monoDiagnosticsMock";
 import { PThreadPtr, Thread } from "../../types/internal";
