+++ src/mono/browser/runtime/diagnostics/mock/environment.ts
@@ -6,7 +6,7 @@ import type { FilterPredicate, MockEnvironment } from "./types";
 import Serializer from "../server_pthread/ipc-protocol/base-serializer";
 import { CommandSetId, EventPipeCommandId, ProcessCommandId } from "../server_pthread/ipc-protocol/types";
 import { assertNever } from "../../types/internal";
-import { pthread_self } from "../../pthreads/worker";
+import { pthread_self } from "../../pthreads";
 import { createPromiseController, mono_assert } from "../../globals";
