+++ src/mono/browser/runtime/exports-internal.ts
@@ -23,7 +23,7 @@ import { mono_wasm_get_func_id_to_name_mappings } from "./logging";
 import { monoStringToStringUnsafe } from "./strings";
 import { mono_wasm_bind_cs_function } from "./invoke-cs";
 
-import { dumpThreads, thread_available } from "./pthreads/browser";
+import { dumpThreads, thread_available } from "./pthreads";
 
 export function export_internal(): any {
     return {
