+++ src/mono/browser/runtime/pthreads/index.ts
@@ -1,0 +1,18 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+export {
+    mono_wasm_main_thread_ptr, mono_wasm_install_js_worker_interop, mono_wasm_uninstall_js_worker_interop,
+    mono_wasm_pthread_ptr, update_thread_info, isMonoThreadMessage, monoThreadInfo,
+} from "./shared";
+export {
+    dumpThreads, thread_available, cancelThreads, is_thread_available,
+    populateEmscriptenPool, mono_wasm_init_threads, init_finalizer_thread,
+    waitForThread, replaceEmscriptenPThreadUI
+} from "./ui-thread";
+export { addUnsettledPromise, settleUnsettledPromise, mono_wasm_eventloop_has_unsettled_interop_promises } from "./worker-eventloop";
+export {
+    mono_wasm_pthread_on_pthread_attached, mono_wasm_pthread_on_pthread_unregistered,
+    mono_wasm_pthread_on_pthread_registered, mono_wasm_pthread_set_name, currentWorkerThreadEvents,
+    dotnetPthreadCreated, initWorkerThreadEvents, replaceEmscriptenPThreadWorker, pthread_self
+} from "./worker-thread";
