+++ src/mono/browser/runtime/pthreads/shared.ts
@@ -4,15 +4,24 @@
 import WasmEnableThreads from "consts:wasmEnableThreads";
 import BuildConfiguration from "consts:configuration";
 
-import { ENVIRONMENT_IS_PTHREAD, Module, loaderHelpers, mono_assert, runtimeHelpers } from "../../globals";
-import { set_thread_prefix } from "../../logging";
-import { bindings_init } from "../../startup";
-import { forceDisposeProxies } from "../../gc-handles";
-import { GCHandle, GCHandleNull, MonoThreadMessage, PThreadPtr, PThreadPtrNull, WorkerToMainMessageType, monoMessageSymbol } from "../../types/internal";
-import { MonoWorkerToMainMessage } from "./types";
-import { monoThreadInfo } from "../worker";
+import type { GCHandle, MonoThreadMessage, PThreadInfo, PThreadPtr } from "../types/internal";
 
-/// Messages sent on the dedicated mono channel between a pthread and the browser thread
+import { ENVIRONMENT_IS_PTHREAD, Module, loaderHelpers, mono_assert, runtimeHelpers } from "../globals";
+import { set_thread_prefix } from "../logging";
+import { bindings_init } from "../startup";
+import { forceDisposeProxies } from "../gc-handles";
+import { monoMessageSymbol, GCHandleNull, PThreadPtrNull, WorkerToMainMessageType } from "../types/internal";
+
+export const worker_empty_prefix = "          -    ";
+
+const monoThreadInfoPartial: Partial<PThreadInfo> = {
+    pthreadId: PThreadPtrNull,
+    reuseCount: 0,
+    updateCount: 0,
+    threadPrefix: worker_empty_prefix,
+    threadName: "emscripten-loaded",
+};
+export const monoThreadInfo: PThreadInfo = monoThreadInfoPartial as PThreadInfo;
 
 export function isMonoThreadMessage(x: unknown): x is MonoThreadMessage {
     if (typeof (x) !== "object" || x === null) {
@@ -103,4 +112,18 @@ export function postMessageToMain(message: MonoWorkerToMainMessage, transfer?: T
     self.postMessage({
         [monoMessageSymbol]: message
     }, transfer ? transfer : []);
+}
+
+export interface MonoWorkerToMainMessage {
+    monoCmd: WorkerToMainMessageType;
+    info: PThreadInfo;
+    port?: MessagePort;
+}
+
+/// Identification of the current thread executing on a worker
+export interface PThreadSelf {
+    info: PThreadInfo;
+    portToBrowser: MessagePort;
+    postMessageToBrowser: <T extends MonoThreadMessage>(message: T, transfer?: Transferable[]) => void;
+    addEventListenerFromBrowser: (listener: <T extends MonoThreadMessage>(event: MessageEvent<T>) => void) => void;
 }
