+++ src/mono/browser/runtime/pthreads/worker-eventloop.ts
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-
 let perThreadUnsettledPromiseCount = 0;
 
 export function addUnsettledPromise() {
