--- src/mono/browser/runtime/pthreads/worker/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "extends": "../../tsconfig.worker.json",
-    "include": [
-        "../../**/*.ts",
-        "../../**/*.d.ts"
-    ]
-}
