+++ src/mono/browser/runtime/scheduling.ts
@@ -5,8 +5,8 @@ import WasmEnableThreads from "consts:wasmEnableThreads";
 
 import cwraps from "./cwraps";
 import { ENVIRONMENT_IS_WORKER, Module, loaderHelpers } from "./globals";
-import { is_thread_available } from "./pthreads/browser/replacements";
 import { forceThreadMemoryViewRefresh } from "./memory";
+import { is_thread_available } from "./pthreads";
 
 let spread_timers_maximum = 0;
 let pump_count = 0;
